backend/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.5
django-cors-headers==4.6.0
django-randomslugfield==0.3.0
djangorestframework==3.15.2
inflection==0.5.1
Markdown==3.7
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
